moved btlx process base class to own module
