+++ src/compas_timber/_fabrication/__init__.py
@@ -1,5 +1,7 @@
 # This module is a tepmorary solution to the problem of circular imports in the compas_timber package.
 # It will be reoved or merged with the `fabrication` module once the migration to the new feature system is complete.
+from .btlx_process import BTLxProcess
 from .jack_cut import JackRafterCut
 
-__all__ = ["JackRafterCut"]
+
+__all__ = ["JackRafterCut", "BTLxProcess"]
